# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Brand Creative Automation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List all custom templates
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a new custom template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a template by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdate"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /templates/dissect-pdf:
    post:
      operationId: dissectPdf
      tags: [templates]
      summary: Dissect an uploaded PDF into a draft freeform template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DissectPdfRequest"
      responses:
        "200":
          description: Draft freeform template extracted from the PDF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DissectPdfResult"
        "400":
          description: Missing or invalid required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: PDF could not be parsed into a template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /templates/dissect-image:
    post:
      operationId: dissectImage
      tags: [templates]
      summary: Learn the layout of an uploaded creative image into a draft freeform template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DissectImageRequest"
      responses:
        "200":
          description: Draft freeform template learned from the image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DissectPdfResult"
        "400":
          description: Missing or invalid required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Image could not be learned into a template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    FreeformElement:
      type: object
      description: A single positioned element in a freeform template layout.
      properties:
        type:
          type: string
          description: "text | image | rect"
        id:
          type: string
        x:
          type: number
        y:
          type: number
        w:
          type: number
        h:
          type: number
        role:
          type: string
        text:
          type: string
        fontSize:
          type: number
        fontWeight:
          type: integer
        color:
          type: string
        align:
          type: string
        lineHeight:
          type: number
        fontFamily:
          type: string
        fontStyle:
          type: string
          description: "normal | italic"
        letterSpacing:
          type: number
        opacity:
          type: number
          description: "0..1"
        src:
          type: ["string", "null"]
        fit:
          type: string
          description: "cover | contain (image objectFit; default depends on role)"
        fill:
          type: string
        radius:
          type: number
        borderColor:
          type: string
        borderWidth:
          type: number
      required:
        - type
    TemplateConfig:
      type: object
      properties:
        contentAlignment:
          type: string
        textAlign:
          type: string
        showAccentBar:
          type: boolean
        showLogoBar:
          type: boolean
        imageStyle:
          type: string
        kind:
          type: string
          description: "preset | freeform (absent means preset)"
        elements:
          type: array
          items:
            $ref: "#/components/schemas/FreeformElement"
    Template:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        width:
          type: integer
        height:
          type: integer
        dims:
          type: string
        config:
          $ref: "#/components/schemas/TemplateConfig"
        sourceImageUrl:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - key
        - name
        - category
        - width
        - height
        - dims
        - config
        - createdAt
        - updatedAt
    TemplateInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        width:
          type: integer
        height:
          type: integer
        config:
          $ref: "#/components/schemas/TemplateConfig"
        sourceImageUrl:
          type: ["string", "null"]
      required:
        - name
        - width
        - height
    TemplateUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        width:
          type: integer
        height:
          type: integer
        config:
          $ref: "#/components/schemas/TemplateConfig"
        sourceImageUrl:
          type: ["string", "null"]
    DissectPdfRequest:
      type: object
      properties:
        objectPath:
          type: string
          minLength: 1
        page:
          type: integer
          minimum: 1
      required:
        - objectPath
    DissectImageRequest:
      type: object
      properties:
        objectPath:
          type: string
          minLength: 1
      required:
        - objectPath
...
    DissectPdfResult:
      type: object
      properties:
        name:
          type: string
        width:
          type: integer
        height:
          type: integer
        config:
          $ref: "#/components/schemas/TemplateConfig"
        warnings:
          type: array
          items:
            type: string
      required:
        - name
        - width
        - height
        - config
        - warnings
...